[Keyboard] add bear_face/v2 (#9642)
* [keyboard] bear_face/v2 initial commit

* restructured keyboard dir for multiple versions
* fixed .json layouts for ISO
* reformatted hard tabs to soft tabs
* updated readmes

* [keyboard] removed top-level info.json, keymaps

* [bear_face/v2] changed matrix height, arrow key labels

* fixed height 6 instead of 5
* symbols for arrows now directions (i.e. "Left" etc.)

* [bear_face/v2] fixed ISO enter in .json

* corrected placement for ISO enter

* Update rules.mk

* DEFAULT_FOLDER = bear_face/v1

* [bear_face] fixed matrix width in info.json

* true matrix width is 15
* updated for v1 and v2 info.json

* Update keyboards/bear_face/config.h

* Update keyboards/bear_face/v1/info.json

* Update keyboards/bear_face/v2/info.json

* Update keyboards/bear_face/v2/info.json

* Update keyboards/bear_face/rules.mk

* Fixed default keymap tabs

* changed from tabs to spaces
