+++ keyboards/bear_face/readme.md
@@ -1,15 +1,16 @@
 # bear_face
 
-![bear_face](https://i.imgur.com/nQIyP6w.png)
+![bear_face](https://i.imgur.com/SALKmdP.png)
 
-Vortex Race 3 with replacement QMK-compatible PCB designed by [chemicalwill](https://github.com/chemicalwill)
+QMK-compatible replacement PCB for the Vortex Race 3 designed by [chemicalwill](https://github.com/chemicalwill)
 
-* Keyboard Maintainer: [Will Hedges](https://github.com/chemicalwill)
-* Hardware Supported: bear_face v1.0, atmega32u4
-* Hardware Availability: [PCB files](https://github.com/chemicalwill/bear_face)
+* Keyboard Maintainer: [chemicalwill](https://github.com/chemicalwill)
+* Hardware Supported: bear_face v1, v2; atmega32u4
+* Hardware Availability: [currently by request](https://www.reddit.com/message/compose?to=chemicalwill&subject=bear_face%20v2)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make bear_face:default
+    make bear_face/v1:default
+    make bear_face/v2:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
