+++ keyboards/bear_face/v1/keymaps/default/readme.md
@@ -5,10 +5,11 @@ This layout replaces the stock layout on the Vortex Race 3.
 - Caps Lock indicator LED is enabled by default
 - Layer Tap on Caps Lock (tap for Caps Lock, hold for _FN1)
 - FORCE_NKRO enabled by default
+
 - Pn key is set to 'KC_NO' by default
-    * might be a good place for a macro, or to put your PC to sleep, etc.
-- a combined function layer that mimics the sublegends on the stock caps (regardless of layout)
+    * Might be a good place for a macro, or to put your PC to sleep, etc.
+
+- A combined function layer that mimics the sublegends on the stock caps (regardless of layout)
     * 'Reset' will put the keyboard into DFU mode
     * 'APP' sends 'KC_APP'
-    * set base layer toggles for QWERTY, COLEMAK, and DVORAK layouts (will persist after reboot)
-    
+    * Base layer toggles for QWERTY, COLEMAK, and DVORAK layouts (will persist after reboot)
