+++ keyboards/bear_face/v2/config.h
@@ -15,4 +15,5 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "bear_face.h"
+/* USB Device descriptor parameter */
+#define DEVICE_VER      0x0200
