+++ keyboards/bear_face/v2/v2.c
@@ -16,3 +16,19 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "bear_face.h"
+
+void keyboard_pre_init_kb(void) {
+    //Sets LED pin as output
+    setPinOutput(F7);
+
+    keyboard_pre_init_user();
+}
+
+bool led_update_kb(led_t led_state) {
+    // Caps Lock LED indicator toggling code here
+    bool res = led_update_user(led_state);
+    if(res) {
+        writePin(F7, led_state.caps_lock);
+    }
+    return res;
+}
